+++ src/commands/moderation/warn/deleteWarn.ts
@@ -1,15 +1,10 @@
 import moment from 'moment';
 import { Colors, TextChannel } from 'discord.js';
 import { CommandDefinition } from '../../../lib/command';
-import { Roles, CommandCategory, Channels } from '../../../constants';
+import { CommandCategory, Channels, RoleGroups } from '../../../constants';
 import { makeEmbed } from '../../../lib/embed';
 import { getConn } from '../../../lib/db';
 import Warn from '../../../lib/schemas/warnSchema';
-
-const permittedRoles = [
-    Roles.ADMIN_TEAM,
-    Roles.MODERATION_TEAM,
-];
 
 const noConnEmbed = makeEmbed({
     title: 'Warn - No Connection',
@@ -43,10 +38,7 @@ const deleteEmbed = makeEmbed({
 
 export const deleteWarn: CommandDefinition = {
     name: ['deletewarn', 'delwarn', 'deletewarning'],
-    requirements: {
-        permissions: ['BanMembers'],
-        roles: permittedRoles,
-    },
+    requirements: { roles: RoleGroups.STAFF },
     description: 'Delete a warning',
     category: CommandCategory.MODERATION,
     executor: async (msg) => {